Rename ActiveSupportBreadcrumbs to Breadcrumbs::ActiveSupportLogger
